(feat) Config for activity stream proxy
We only have the dev environment for pen testing. We don't actually plan to use
this in production: it's a "just in case" project, jfor pen testing now, so we
can quickly use a stand alone endpoint down the line without having to wait+pay
for another round of pen-testing.