Add `APPLICATION_ENV` for People Finder and Workspace
Errors from People Finder and Digital Workspace show up as coming from
the `production` environment in Sentry, regardless of whether they come
from dev, staging, or prod. This is because Raven picks up on the Rails
environment by default, which (correctly) is `production` for all three.

This adds an environment variable `APPLICATION_ENV` which we can use in
the applications to configure Raven to report the environment the apps
are actually running in.