Remove config for data-flow-worker-tensorflow
This hasn't been used since around
https://github.com/uktrade/data-flow/pull/2307 when the separate tensorflow
worker was moved into the regular data-flow worker setup, using Airflow's
virtual environment operator to install tensorflow when needed.

This is being done now because we're about to add some more "apps" as part of
the data-flow setup, so doing a small cleanup ahead of time.