Remove dev-a environment from data-flow
The dev-a environment of data-flow no longer exists, and hasn't for a while, so
there isn't a need for its config file.

This is being done because we're about to add some more apps to the data-flow
setup, so cleaning house a bit first so that setup is as clear as it can be.