Add data-flow-data-infrastructure config
data-flow is being split into multiple GOV.UK PaaS apps

- One "core" one for the Airflow webserver and Airflow scheduler
- Then one per team of people that write and maintain DAGs that will run
  Airflow's DAG processor, and run the Airflow tasks themselves.

This is primarily to separate the credentials for the various components, so
everything doesn't get access to everything. The webserver and scheduler don't
need credentials to access data, and also each team's DAGs don't need access to
the credentials for other teams' DAGs.

Since Data Infrastructure is de-facto responsible for all the DAGs, this is the
first team chosen to split out, hence the name of this application.